generate cgo Objective-C bindings
Generates cgo code and Go wrappers to call the Objective-C methods. Type information in `api` is generated using `macschema` to fetch and parse from the docs.

`go run ./gen/cmd` generates the `*_objc.gen.go` files containing the cgo code to invoke each Objective-C method, and Go wrappers on top of it.

The generated types are meant to be embedded by the author in a struct to include the generated methods, but allow the author to augment them with additional methods, like:
```
type NSFoo struct { gen_NSFoo }
```